% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genBin.R
\name{genBin}
\alias{genBin}
\title{Generate the correlated binary response data for discrete case}
\usage{
genBin(corstr, mu = NULL, size = NULL)
}
\arguments{
\item{corstr}{A character string specifying the correlation structure for the
clusters. Allowed structures are: "independence", "exchangeable" and "ar1".}

\item{mu}{A numeric parameter denotes the value of the link function}

\item{size}{A numeric number indicating the size of the matrix.}
}
\value{
a function to get the correlated binary response data
}
\description{
\code{genBin} generate the data used for discrete case
}
\details{
genBin returns the correlated binary response according to the value of the
corstr argument.
}
