% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCorMat.R
\name{genCorMat}
\alias{genCorMat}
\title{Generate the correlation matrix for the clusteded data}
\usage{
genCorMat(corstr, rho, size)
}
\arguments{
\item{corstr}{A character string specifying the correlation structure for the
clusters. Allowed structures are: "independence", "exchangeable" and "ar1".}

\item{rho}{A numeric parameter in correlation structure for the
autocorrelation coefficient.}

\item{size}{A numeric number indicating the size of the matrix.}
}
\value{
a matrix which represents the diffetent correlation matrix.
}
\description{
\code{genCorMat} generate the data with specified correlation matrix.
}
\details{
genCorMat returns the corresponding correlation matrix according to the value
of the corstr argument.
}
