% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_stop_ASE.R
\name{is_stop_ASE}
\alias{is_stop_ASE}
\title{Determining whether to stop choosing sample}
\usage{
is_stop_ASE(sandwich, d, nonZeroIdx, verbose = FALSE)
}
\arguments{
\item{sandwich}{A numeric matrix that represent the sandwich information
matrix for covariance}

\item{d}{A numeric number specifying the length of the fixed size confidence
set that we specify}

\item{nonZeroIdx}{A numeric number specifying the index of the non zero
coefficient}

\item{verbose}{A A logical value to determine whether to get the full
  information about the iteration
situation}
}
\value{
a list of these elements:
\item{stop}{a logical value. If TRUE, it means we have choosen enough
samples.}
\item{eigen}{the maximum eighevalue covariance matrix}
}
\description{
\code{is_stop_ASE} determines whether to stop choosing sample based on the
current estimator
}
\details{
is_stop_ASE determines if the iteration stop condition is met based on the
current estimator
}
