% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_bin.R
\name{print.seqbin}
\alias{print.seqbin}
\title{Print the results by the binary logistic regression model}
\usage{
\method{print}{seqbin}(x, ...)
}
\arguments{
\item{x}{A variable of type \code{seqbin}}

\item{...}{Additional variables to be transferred to functions}
}
\value{
print.seqbin returns several messages to show the conclusion.
}
\description{
\code{print.seqbin} print the result of the binary logistic regression model
used by the method of adaptive shrinkage estimate.
}
\details{
This function is used to present results in a concise way. If we select
enough samples that satisfy the stopping criterion, then we show several
messages to report the conclusion including the length of fixed size
confidence set, the number of samples we choose, the value of coefficient and
the time have elapsed. Otherwise, the sample selection process is failed. We
need to reduce the length of fixed size confidence set because the smaller
the dlen, the larger the sample size we need.
}
