% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_multi.R
\name{print.seqmulti}
\alias{print.seqmulti}
\title{Print the results by the multi-logistic regression model}
\usage{
\method{print}{seqmulti}(x, ...)
}
\arguments{
\item{x}{A variable of type \code{seqmulti}}

\item{...}{Additional variables to be transferred to functions}
}
\value{
print.seqmulti returns several messages to show the conclusion.
}
\description{
\code{print.seqmulti} print the result of the multi-logistic regression model
}
\details{
This function is used to present results in a concise way. If we select
enough samples that satisfy the stopping criterion, then we show several
messages to report the conclusion including the length of fixed size
confidence set, the number of samples we choose and the value of coefficient.
}
